Add logstash-filter-prune 0.1.5
Abandon the idea of using gitfat to ship raw gems and instead just do
the simplest thing possible which is to unpack the needed gem into this
repo and then configure logstash to source it using the `--pluginpath`
directive. Thankfully upstream came to their senses and restored this
functionality in 1.5.3 after removing it in 1.5.0
(<https://github.com/elastic/logstash/pull/3587>).

Bug: T99735
Change-Id: Ied616e2e9353567cd2aaffbf57b5114620dcc24f