Add de_dot filter and rename to logstash-filters-wikimedia
To facilitate the upgrade to elasticsearch 2.x we need to add the de_dot
filter, as 2.x does not allow properties to contain a dot.  Additionally
renames the plugin to logstash-filters-wikimedia to capture the fact
that this now contains multiple plugins and not just the prune plugin.

The de_dot filter is unchanged from upstream. Although it is marked as
requiring logstash >= 2.0.0 it works just fine against 1.5.3.

Bug: T138335
Change-Id: Ibb835587d90a5483e8d38b2481bae5c9eae4fd83