Update logstash plugins for 5.x
This creates a plugin pack because logstash 5.x no longer has a
catch-all contribs that includes everything under the sun. Additionally
the `plugin.path` configuration that was previously used to point to our
custom plugins has been deprecated in 5.x and will be removed in 6.x.
Of the two methods provided in the logstash documentation (plugin pack
or setup and run a private gem repository) this seems like the least
maintenance intensive manner.

* Moves custom filters into subdir to facilitate building a gem
* Removes the prune plugin in favor of upstream logstash-filter-prune
* Adds a build.sh script to automate building a logstash plugin pack.

Change-Id: I162014dc50f0d5930dfddd9391b8f0cb046214ba